Replace colors with chalk due to malicious version of colors
The maintainer of colors has purposely introduced an infinite loop
in the latest version of colors. Therefore we need to ditch this
dependency.

cli-table must also be updated in order to get rid of the colors
dependency entirely. There is a PR:
https://github.com/Automattic/cli-table/pull/166
Or maybe an alternative table lib should be considered as well?