Use release date of latest version instead of modified date
The `time.modified` field is apparently not the time of the latest
release, it can be much more recent. I have not found any good
documentation on exactly what "modified" means.

I think using the most recent release date is the best measure for
detecting if a package is ditched or not. Note that this does not mean
the version tagged with the "latest" tag - this will also consider other
tags, like rc, alpha, next, etc.